fix architecture to be more restful
