+++ back/routes/index.js
@@ -2,18 +2,10 @@ const express = require("express");
 const router = express.Router();
 const userRouter = require("./users");
 const bookRouter = require("./books");
-const User = require("../models/User");
-const passport = require("../config/passport");
+const sessionRouter = require("./sessions");
 
 router.use("/users", userRouter);
 router.use("/books", bookRouter);
-
-router.post("/register", function(req, res) {
-  User.create(req.body)
-    .then(user => {
-      res.status(201).send(user);
-    })
-    .catch(err => res.status(400).send(console.log(err)));
-});
+router.use("/sessions", sessionRouter);
 
 module.exports = router;
