+++ back/routes/users.js
@@ -1,23 +1,13 @@
 const express = require("express");
 const router = express.Router();
-const passport = require("../config/passport");
+const User = require("../models/User");
 
-function isLogedIn(req, res, next) {
-  if (req.isAuthenticated()) {
-    res.send(req.user.email);
-  } else {
-    res.send(false);
-  }
-}
-
-router.post("/login", passport.authenticate("local"), (req, res) => {
-  if (req.isAuthenticated()) {
-    res.json(req.user.dataValues.name);
-  } else {
-    res.status(401).res.json({});
-  }
+router.post("/", function(req, res) {
+  User.create(req.body)
+    .then(user => {
+      res.status(201).send(user);
+    })
+    .catch(err => res.status(400).send(console.log(err)));
 });
-
-router.get("/", isLogedIn);
 
 module.exports = router;
