+++ front/src/store/actions/users.js
@@ -12,7 +12,7 @@ const logUser = logUser => ({
    });
 
 export const signUpUser = (user) => dispatch =>{
-    return axios.post('register', user)
+    return axios.post('/users', user)
     .then(user => true)
     .catch(err => {
         throw err
@@ -24,7 +24,7 @@ export const fetchUser = () => dispatch =>
         .then(user => dispatch(getUser(user.data)))
 
 export const loginUser = (username, password) => dispatch =>{
-    return axios.post("/users/login", {username, password})
+    return axios.post("/sessions", {username, password})
         .then(res => dispatch(logUser(res.data)))
         .catch(err => {
             throw err
